requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-admin-interface==0.28.8
django-ckeditor==6.7.1
django-colorfield==0.11.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-recaptcha==4.0.0
gevent==24.2.1
greenlet==3.0.3
gunicorn==23.0.0
idna==3.7
packaging==24.1
pillow==10.4.0
python-slugify==8.0.4
requests==2.32.3
setuptools==72.1.0
sqlparse==0.5.1
text-unidecode==1.3
urllib3==2.2.2
zope.event==5.0
zope.interface==7.0.1
